--- Solutions/CipherCloud/Data Connectors/CipherCloudConnector/CipherCloudConnector/function.json
@@ -1,11 +1,0 @@
-{
-  "scriptFile": "__init__.py",
-  "bindings": [
-    {
-      "name": "mytimer",
-      "type": "timerTrigger",
-      "direction": "in",
-      "schedule": "0 */5 * * * *"
-    }
-  ]
-}
