--- Solutions/CipherCloud/Data Connectors/CipherCloudConnector/requirements.txt
@@ -1,10 +1,0 @@
-# DO NOT include azure-functions-worker in this file
-# The Python Worker is managed by Azure Functions platform
-# Manually managing azure-functions-worker may cause unexpected issues
-
-pyjwt
-azure-functions
-azure-storage-file-share==12.3.0
-python-dateutil==2.8.1
-azure_core==2.0.12
-requests==2.25.1
