+++ Solutions/Lookout Cloud Security/Data Connectors/LookoutCSConnector/proxies.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "http://json.schemastore.org/proxies",
+  "proxies": {}
+}
