+++ Solutions/Lookout Cloud Security/Parsers/LookoutCSActivities.txt
@@ -1,0 +1,23 @@
+// Usage Instruction :
+// Paste below query in log analytics, click on Save button and select as Function from drop down by specifying function name and alias as CipherCloudEvents.
+// Function usually takes 10-15 minutes to activate. You can then use function alias from any other queries (e.g. LookoutCSActivities | take 10).
+// Reference : Using functions in Azure monitor log queries : https://docs.microsoft.com/azure/azure-monitor/log-query/functions
+Lookoutlogs_CL
+| extend timeStamp = timeStamp_t
+| extend eventType = eventType_s
+| extend actionType = actionType_s
+| extend activityType = activityType_s
+| extend appName = appName_s
+| extend contentName = contentName_s
+| extend contentUrl = column_ifexists('contentUrl_s', '')
+| extend eventId = eventId_g
+| where eventType == 'Activity'
+| project
+    timeStamp,
+    eventType,
+    actionType,
+    activityType,
+    appName,
+    contentName,
+    contentUrl,
+    eventId
