+++ Solutions/Lookout Cloud Security/Parsers/LookoutCSAnomalies.txt
@@ -1,0 +1,31 @@
+// Usage Instruction :
+// Paste below query in log analytics, click on Save button and select as Function from drop down by specifying function name and alias as CipherCloudEvents.
+// Function usually takes 10-15 minutes to activate. You can then use function alias from any other queries (e.g. LookoutCSAnomalies | take 10).
+// Reference : Using functions in Azure monitor log queries : https://docs.microsoft.com/azure/azure-monitor/log-query/functions
+Lookoutlogs_CL
+| extend timeStamp = timeStamp_t
+| extend eventType = eventType_s
+| extend actionType = actionType_s
+| extend anomalyType = column_ifexists('anomalyType_s', '')
+| extend appName = appName_s
+| extend userEmail = userEmail_s
+| extend anomalyName = column_ifexists('anomalyName_s', '')
+| extend currentCity = column_ifexists('currentCity_s', '')
+| extend currentTimestamp = column_ifexists('currentTimestamp_t', '')
+| extend previousTimestamp = column_ifexists('previousTimestamp_t', '')
+| extend currentEventId = column_ifexists('currentEventId_g', '')
+| extend previousEventId = column_ifexists('previousEventId_g', '')
+| where eventType == 'Anomaly'
+| project
+    timeStamp,
+    eventType,
+    actionType,
+    anomalyType,
+    appName,
+    userEmail,
+    anomalyName,
+    currentCity,
+    currentTimestamp,
+    previousTimestamp,
+    currentEventId,
+    previousEventId
