+++ Solutions/Lookout Cloud Security/Parsers/LookoutCSViolations.txt
@@ -1,0 +1,34 @@
+// Usage Instruction :
+// Paste below query in log analytics, click on Save button and select as Function from drop down by specifying function name and alias as CipherCloudEvents.
+// Function usually takes 10-15 minutes to activate. You can then use function alias from any other queries (e.g. LookoutCSViolations | take 10).
+// Reference : Using functions in Azure monitor log queries : https://docs.microsoft.com/azure/azure-monitor/log-query/functions
+Lookoutlogs_CL
+| extend timeStamp = timeStamp_t
+| extend eventType = eventType_s
+| extend actionType = actionType_s
+| extend activityType = activityType_s
+| extend appName = appName_s
+| extend cloudType = cloudType_s
+| extend contentName = contentName_s
+| extend contentUrl = column_ifexists('contentUrl_s', '')
+| extend eventId = eventId_g
+| extend externalCollaborators = column_ifexists('externalCollaborators_s', '')
+| extend policyName = policyName_s
+| extend scanType = column_ifexists('scanType_s', '')
+| extend userEmail = userEmail_s
+| extend violation = violation_s
+| where eventType == 'Violation'
+| project timeStamp,
+    eventType, 
+    actionType,
+    activityType, 
+    appName,
+    cloudType,
+    contentName,
+    contentUrl, 
+    eventId,
+    externalCollaborators, 
+    policyName,
+    scanType,
+    userEmail, 
+    violation
